+++ packages/components/lib/Atoms/Icon/Icon.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import Icon from 'svelte-awesome/components/Icon.svelte'
+	export let icon
+</script>
+
+<Icon data={icon} />