+++ packages/components/lib/Atoms/MenuItem/MenuItem.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+	export let href: string
+</script>
+
+<li>
+	<a class="pl-1" {href}>
+		<slot />
+	</a>
+</li>