+++ packages/metadata/CHANGELOG.md
@@ -13,5 +13,5 @@
 - # Updates
   - updated sveltekit to 1.0.0-next.405
   - updated "plop app" to scaffold sveltekit app
-  - renamed @packages/ui to @devmoon/ui
+  - renamed @packages/ui to @devmoon/components
   - added a bunch of new components