+++ packages/ui/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @devmoon/ui
+# @devmoon/components
 
 ## 0.2.0
 
@@ -24,7 +24,7 @@
 - # Updates
   - updated sveltekit to 1.0.0-next.405
   - updated "plop app" to scaffold sveltekit app
-  - renamed @packages/ui to @devmoon/ui
+  - renamed @packages/ui to @devmoon/components
   - added a bunch of new components
 
 ### Patch Changes